refactor(GameManager): 자동차 등록 기능 수정
- set함수에 인자를 추가
- 이는 의존성을 약화시키고 테스트 코드 작성이 용이하게 하기 위함임
